+++ code/FINDER_ND_cost/nstep_replay_mem_prioritized.pyx
@@ -39,8 +39,8 @@ cdef class py_Data:
         num_nodes = graph1.num_nodes     #得到Graph对象的节点个数
         num_edges = graph1.num_edges    #得到Graph对象的连边个数
         edge_list = graph1.edge_list
-        cint_edges_from = np.zeros([num_edges],dtype=np.int)
-        cint_edges_to = np.zeros([num_edges],dtype=np.int)
+        cint_edges_from = np.zeros([num_edges],dtype=np.int32)
+        cint_edges_to = np.zeros([num_edges],dtype=np.int32)
         for i in range(num_edges):
             cint_edges_from[i]=edge_list[i].first
             cint_edges_to[i] =edge_list[i].second
@@ -165,8 +165,8 @@ cdef class py_ReplaySample:
         edge_list = graph1.edge_list
         nodes_weight = graph1.nodes_weight
 
-        cint_edges_from = np.zeros([num_edges],dtype=np.int)
-        cint_edges_to = np.zeros([num_edges],dtype=np.int)
+        cint_edges_from = np.zeros([num_edges],dtype=np.int32)
+        cint_edges_to = np.zeros([num_edges],dtype=np.int32)
         cdouble_nodes_weight = np.zeros([num_nodes],dtype=np.double)
 
         cdef int i
